Move core functions to core.py
