+++ src/wam2layers/preprocessing/shared.py
@@ -5,40 +5,6 @@ import pandas as pd
 import xarray as xr
 from scipy.interpolate import interp1d
 
-from wam2layers.config import Config
-from wam2layers.utils.profiling import ProgressTracker
-
-
-# old, keep only for reference and ecearth starting case
-def resample(variable, divt, count_time, method="interp"):
-    """Resample the variable to a given number of timesteps."""
-
-    _, nlat, nlon = variable.shape
-    ntime = count_time * divt
-    shape = (ntime, nlat, nlon)
-    new_var = np.nan * np.zeros(shape)
-
-    oddvector = np.zeros(ntime)
-    partvector = np.zeros(ntime)
-    for o in np.arange(0, ntime, divt):
-        for i in range(1, divt):
-            oddvector[o + i - 1] = (divt - i) / divt
-            partvector[o + i] = i / divt
-
-    for t in range(ntime):
-        idx = int((t + 1) / divt + oddvector[t])
-
-        if method == "bfill":
-            new_var[t] = (1 / divt) * variable[idx - 1]
-        elif method == "interp":
-            new_var[t] = variable[idx - 1] + partvector[t] * (
-                variable[idx] - variable[idx - 1]
-            )
-        else:
-            raise ValueError(f"Unknown resample method {method}")
-
-    return new_var
-
 
 def join_levels(pressure_level_data, surface_level_data):
     """Combine 3d pressure level and 2d surface level data.
@@ -256,122 +222,3 @@ def accumulation_to_flux(data, input_frequency):
     # Extrapolation introduces a small inconsistency at the last midnight...
     # data.interp(time=original_time, kwargs={"fill_value": "extrapolate"})
     return fluxdata
-
-
-def stabilize_fluxes(F, S, progress_tracker: ProgressTracker, config: Config, t):
-    """Stabilize the outfluxes / influxes.
-
-    CFL: Water cannot move further than one grid cell per timestep.
-    """
-    for level in ["upper", "lower"]:
-        fx = F["fx_" + level] * config.timestep
-        fy = F["fy_" + level] * config.timestep
-        s = S["s_" + level]
-
-        fx_abs = np.abs(fx)
-        fy_abs = np.abs(fy)
-        ft_abs = fx_abs + fy_abs
-
-        # TODO: make 1/2 configurable?
-        fx_limit = 1 / 2 * fx_abs / ft_abs * s.values
-        fx_stable = np.minimum(fx_abs, fx_limit)
-
-        fy_limit = 1 / 2 * fy_abs / ft_abs * s.values
-        fy_stable = np.minimum(fy_abs, fy_limit)
-
-        progress_tracker.track_stability_correction(fy_stable, fy_abs, config, t)
-
-        # Get rid of any nan values
-        fx_stable.fillna(0)
-        fy_stable.fillna(0)
-
-        # Re-instate the sign and convert back to flux instead of accumulation
-        F["fx_" + level] = np.sign(fx) * fx_stable / config.timestep
-        F["fy_" + level] = np.sign(fy) * fy_stable / config.timestep
-
-
-def divergence(fx, fy):
-    # Note: latitude decreasing, hence negative fy gradient is divergence
-    return np.gradient(fx, axis=-1) - np.gradient(fy, axis=-2)
-
-
-def calculate_fz(F, S0, S1, dt, kvf):
-    """Calculate the vertical fluxes.
-
-    The vertical flux is calculated as a closure term. Residuals are distributed
-    proportionally to the amount of moisture in each layer.
-
-    The flux is constrained such that it can never exceed
-
-    Arguments:
-        F: xarray dataset with fluxes evaluated at temporal midpoints between states
-        S0: xarray dataset with states at current time t
-        S1: xarray dataset with states at updated time t+1 (always forward looking)
-        dt: timestep
-        kvf: net to gross vertical flux multiplication parameter. With the a value of 0,
-            the net and gross fluxes are equal.
-
-    Returns:
-        fz: vertical flux, positive downward
-
-    Examples:
-
-        Create dummy input data. In the absence of any horizontal fluxes and
-        sources or sinks, this result can be verified manually.
-
-        >>> import numpy as np
-        >>> import xarray as xr
-        >>> from wam2layers.preprocessing.shared import calculate_fz
-        >>>
-        >>> F = xr.Dataset({
-        ...     'precip': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ...     'evap': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ...     'fx_upper': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ...     'fy_upper': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ...     'fx_lower': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ...     'fy_lower': xr.DataArray(np.zeros(((3, 3))), dims=['lat', 'lon']),
-        ... })
-        >>> S0 = xr.Dataset({
-        ...     's_upper': 10 * xr.DataArray(np.ones(((3, 3))), dims=['lat', 'lon']),
-        ...     's_lower': 6 * xr.DataArray(np.ones(((3, 3))), dims=['lat', 'lon']),
-        ... })
-        >>> S1 = xr.Dataset({
-        ...     's_upper': 8 * xr.DataArray(np.ones(((3, 3))), dims=['lat', 'lon']),
-        ...     's_lower': 7 * xr.DataArray(np.ones(((3, 3))), dims=['lat', 'lon']),
-        ... })
-        >>> calculate_fz(F, S0, S1, dt=1)
-        <xarray.DataArray (lat: 3, lon: 3)>
-        array([[1.46666667, 1.46666667, 1.46666667],
-               [1.46666667, 1.46666667, 1.46666667],
-               [1.46666667, 1.46666667, 1.46666667]])
-        Dimensions without coordinates: lat, lon
-
-    """
-    s_mean = (S1 + S0) / 2
-    s_total = s_mean.s_upper + s_mean.s_lower
-    s_rel = s_mean / s_total
-
-    # Evaluate all terms in the moisture balance execpt the unknown Fz and err
-    foo_upper = (S1 - S0).s_upper + dt * (
-        divergence(F.fx_upper, F.fy_upper) + F.precip.values * s_rel.s_upper
-    )
-    foo_lower = (S1 - S0).s_lower + dt * (
-        divergence(F.fx_lower, F.fy_lower) + F.precip.values * s_rel.s_lower - F.evap
-    )
-
-    # compute the resulting vertical moisture flux; the vertical velocity so
-    # that the new err_lower/s_lower = err_upper/s_upper (positive downward)
-    fz = foo_lower - S1.s_lower / (S1.s_lower + S1.s_upper) * (foo_upper + foo_lower)
-
-    # TODO: verify that err_lower/s_lower = err_upper/s_upper is satisfied on debug log
-
-    # stabilize the outfluxes / influxes; during the reduced timestep the
-    # vertical flux can maximally empty/fill 1/x of the top or down storage
-    stab = 1.0 / (kvf + 1.0)
-    flux_limit = np.minimum(
-        s_mean.s_upper, s_mean.s_lower
-    )  # TODO why is this not 'just' the upstream bucket?
-    fz_stable = np.minimum(np.abs(fz), stab * flux_limit)
-
-    # Reinstate the sign and convert accumulation back to flux
-    return np.sign(fz) * fz_stable / dt
